#7 Directly implement Terrain types
