+++ VpNet/ManagedApi/Scene/Terrain/Enumerations/TerrainRotation.cs
@@ -1,5 +1,8 @@
 namespace VpNet
 {
+    /// <summary>
+    ///     An enumeration of terrain rotations.
+    /// </summary>
     public enum TerrainRotation
     {
         North,
